+++ changelog.md
@@ -6,10 +6,13 @@
 
 ### Changed
 
+- When updating configuration, if collection or field name is customized, keep the customized name.
+
 ### Fixed
 
 - Predicates in relationships using in ordering (usually supplied by the engine's permission system) would fail to join the related tables correctly if the predicate was null.
   [#655](https://github.com/hasura/ndc-postgres/pull/655)
+- Table names that conflict with scalar type names will now be aliased, with a suffix starting with `_table`, and from then `_table_n` where `n` is an incrementing integer, until a unique name is found. [hasura/graphql-engine#10570](https://github.com/hasura/graphql-engine/issues/10570)
 
 ## [v1.2.0] - 2024-10-25